Have kafka:info display config vars from attachments
The config var information in kafka:info is based on metadata from the
provider endpoint in the kafka API. However, that may not be there
until a while after the initial provision, during which period the
config vars show up empty (or, prior to the refactor, incorrectly
defaulting to `KAFKA_URL`).

However, the information can instead be fetched from the
`/addon-attachments` Heroku API endpoint. This does not work for
"pseudo-attachments" (i.e., just copying around config vars), but
that's strongly discouraged anyway and we shouldn't support it.