Fix kafka:tail for v6 compatibility (#161)
Have tail return a promise; resolve on SIGINT

This lets us be compatible with v6 CLI.

Also change some exits to throw instead and let the CLI handle errors,
and update tests to work with the new mechanism. The changes require
us to mock out `process` in tests, so we export it from the command
and let our tests override it.