Make kafka:credentials --reset flag required
The CLI framework has some quirks around boolean flags, including that
they don't obey the required option (which admittedly is kind of a
weird thing to do; this was originally done for consistency with
Postgres but now Postgres does something different). Work around this
by explicitly checking for the presence of the flag.